+++ kirbycms-extension-image-lib-gd.php
@@ -129,16 +129,25 @@ class ImageExtDriverGD extends ImageExtDriver {
 	}
 	
 	public function best_fit($max_width, $max_height, $upscale = false){
+		if ( $upscale !== false && ( $width >= $this->image_class->get_width() || $height >= $this->image_class->get_height() ) ){
+			return $this;
+		}
 		$this->image_class->best_fit($max_width, $max_height);
 		return $this;
 	}
 	
 	public function fit_to_height($height, $upscale = false){
+		if ( $upscale !== false && $height >= $this->image_class->get_height() ){
+			return $this;
+		}
 		$this->image_class->fit_to_height($height);
 		return $this;
 	}
 	
 	public function fit_to_width($width, $upscale = false){
+		if ( $upscale !== false && $width >= $this->image_class->get_width() ){
+			return $this;
+		}
 		$this->image_class->fit_to_width($width);
 		return $this;
 	}
@@ -159,6 +168,9 @@ class ImageExtDriverGD extends ImageExtDriver {
 	}
 	
 	public function resize($width, $height, $upscale = false){
+		if ( $upscale !== false && ( $width >= $this->image_class->get_width() || $height >= $this->image_class->get_height() ) ){
+			return $this;
+		}
 		$this->image_class->resize($width, $height);
 		return $this;
 	}
@@ -169,6 +181,9 @@ class ImageExtDriverGD extends ImageExtDriver {
 	}
 	
 	public function thumbnail($width, $height = null, $upscale = false){
+		if ( $upscale !== false && ( $width >= $this->image_class->get_width() || $height >= $this->image_class->get_height() ) ){
+			return $this;
+		}
 		$this->image_class->thumbnail($width, $height);
 		return $this;
 	}